Moved common type definitions from 'src' to 'src/client', so that 'module_dist' can build with a flat folder structure.
