+++ src/provider/workspaces/restore.ts
@@ -2,7 +2,7 @@ import {Application} from 'hadouken-js-adapter/out/types/src/api/application/app
 import {Identity} from 'hadouken-js-adapter/out/types/src/identity';
 
 import {promiseMap} from '../snapanddock/utils/async';
-import {Layout, LayoutApp, LayoutName, WindowState} from '../../types';
+import {Layout, LayoutApp, LayoutName, WindowState} from '../../client/types';
 
 import {regroupLayout} from './group';
 import {providerChannel} from '../main';
